Refactor to use a single app instead of gerry-rigging two
