--- client/package.json
@@ -1,68 +1,0 @@
-{
-  "name": "omega",
-  "version": "0.1.0",
-  "private": true,
-  "license": "MIT",
-  "dependencies": {
-    "@craco/craco": "^6.1.1",
-    "@fortawesome/fontawesome-svg-core": "^1.2.34",
-    "@fortawesome/free-solid-svg-icons": "^5.15.2",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/uuid": "^8.3.0",
-    "fp-ts": "^2.9.5",
-    "io-ts": "^2.2.14",
-    "io-ts-types": "^0.5.13",
-    "monocle-ts": "^2.3.10",
-    "newtype-ts": "^0.3.4",
-    "normalize.css": "^8.0.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-markdown": "^6.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.2",
-    "typescript": "^4.1.2",
-    "uuid": "^8.3.2",
-    "web-vitals": "^1.0.1"
-  },
-  "scripts": {
-    "start": "craco start",
-    "build": "craco build",
-    "test": "craco test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "proxy": "http://localhost:5000/",
-  "devDependencies": {
-    "@tailwindcss/postcss7-compat": "^2.1.0",
-    "@types/react-test-renderer": "^17.0.1",
-    "autoprefixer": "9",
-    "postcss": "7",
-    "react-test-renderer": "^17.0.1",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat"
-  }
-}
