--- jest.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-    preset: 'ts-jest',
-    testEnvironment: 'node',
-    testPathIgnorePatterns: [
-        "<rootDir>/node_modules/",
-        "<rootDir>/client/"
-    ]
-};
