+++ package.json
@@ -6,30 +6,77 @@
   "engines": {
     "node": "16.x"
   },
-  "scripts": {
-    "client": "cd client && yarn start",
-    "server": "npx nodemon server.js",
-    "dev": "concurrently --kill-others-on-fail \"yarn server\" \"yarn client\"",
-    "start": "node server/src/index.js",
-    "build": "tsc",
-    "test": "jest",
-    "heroku-postbuild": "echo 'Nothing to do'"
-  },
   "dependencies": {
-    "@tsconfig/node12": "^1.0.7",
+    "@craco/craco": "^6.1.1",
+    "@fortawesome/fontawesome-svg-core": "^1.2.34",
+    "@fortawesome/free-solid-svg-icons": "^5.15.2",
+    "@fortawesome/react-fontawesome": "^0.1.14",
+    "@tsconfig/node16": "^1.0.1",
     "@types/express": "^4.17.11",
+    "@types/jest": "^26.0.15",
+    "@types/node": "^12.0.0",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
+    "@types/react-router-dom": "^5.1.7",
+    "@types/uuid": "^8.3.0",
     "body-parser": "^1.18.3",
     "express": "^4.16.4",
     "fp-ts": "^2.9.5",
     "io-ts": "^2.2.14",
     "io-ts-types": "^0.5.13",
-    "typescript": "^4.1.5"
+    "monocle-ts": "^2.3.10",
+    "newtype-ts": "^0.3.4",
+    "normalize.css": "^8.0.1",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-markdown": "^6.0.2",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "4.0.2",
+    "typescript": "^4.1.2",
+    "uuid": "^8.3.2",
+    "web-vitals": "^1.0.1"
   },
+  "scripts": {
+    "build": "concurrently --kill-others-on-fail \"yarn build-server\" \"yarn build-client\"",
+    "build-client": "craco build",
+    "build-server": "tsc --project tsconfig.server.json",
+    "client": "craco start",
+    "dev": "concurrently --kill-others-on-fail \"yarn server\" \"yarn client\"",
+    "eject": "react-scripts eject",
+    "heroku-postbuild": "echo 'Nothing to do'",
+    "server": "npx nodemon server.js",
+    "start": "node build-server/server/index.js",
+    "test": "craco test"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "proxy": "http://localhost:5000/",
   "devDependencies": {
-    "@types/jest": "^26.0.24",
-    "@types/node": "^14.14.26",
+    "@tailwindcss/postcss7-compat": "^2.1.0",
+    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/react": "^11.1.0",
+    "@testing-library/user-event": "^12.1.10",
+    "@types/react-test-renderer": "^17.0.1",
+    "autoprefixer": "9",
     "concurrently": "^4.0.1",
-    "jest": "26.6.0",
-    "ts-jest": "^26.5.6"
+    "postcss": "7",
+    "react-test-renderer": "^17.0.1",
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat"
   }
 }
