+++ tsconfig.json
@@ -1,22 +1,29 @@
 {
-    "extends": "@tsconfig/node12/tsconfig.json",
-    "$schema": "https://json.schemastore.org/tsconfig",
-    "display": "Node 12",
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "es2019",
-        "strict": true,
-        "esModuleInterop": true,
-        "skipLibCheck": true,
-        "forceConsistentCasingInFileNames": true,
-        "allowSyntheticDefaultImports": true,
-        "outDir": "./server",
-        "moduleResolution": "node",
-        "sourceMap": true,
-        "resolveJsonModule": true
-    },
-    "compileOnSave": true,
-    "include": [
-        "./src/"
-    ]
+  "compilerOptions": {
+    "target": "es5",
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
+    "allowJs": true,
+    "skipLibCheck": true,
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "strict": true,
+    "forceConsistentCasingInFileNames": true,
+    "noFallthroughCasesInSwitch": true,
+    "module": "esnext",
+    "moduleResolution": "node",
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "noEmit": true,
+    "jsx": "react-jsx"
+  },
+  "include": [
+    "src"
+  ],
+  "exclude": [
+    "src/server"
+  ]
 }
