+++ tsconfig.server.json
@@ -1,0 +1,22 @@
+{
+    "extends": "@tsconfig/node16/tsconfig.json",
+    "$schema": "https://json.schemastore.org/tsconfig",
+    "display": "Node 16",
+    "compilerOptions": {
+        "module": "commonjs",
+        "target": "es2019",
+        "strict": true,
+        "esModuleInterop": true,
+        "skipLibCheck": true,
+        "forceConsistentCasingInFileNames": true,
+        "allowSyntheticDefaultImports": true,
+        "outDir": "./build-server",
+        "moduleResolution": "node",
+        "sourceMap": true,
+        "resolveJsonModule": true
+    },
+    "compileOnSave": true,
+    "include": [
+        "./src/server/"
+    ]
+}
